Remove TransportCreateIndexAction
5.10 added dedicated actions for create-blob and resize - which were the
main users of the create-index action outside of tests.

Removing the create-index logic will allow to specialize the remaining
`createIndex` logic in `MetadataCreateIndexService` for the resize case
- which will make the introduction of `RelationMetadata.Table` easier.

Downside is that this has to remove a few test cases related to orphaned
indices. I plan to bring these back in some form if they're still
relevant after the `RelationMetadata.Table` integration.
