+++ server/src/main/java/io/crate/planner/node/ddl/CreateBlobTablePlan.java
@@ -26,9 +26,6 @@ import static io.crate.blob.v2.BlobIndicesService.SETTING_INDEX_BLOBS_ENABLED;
 import java.util.function.Function;
 
 import org.elasticsearch.Version;
-import org.elasticsearch.action.admin.indices.create.CreateIndexAction;
-import org.elasticsearch.action.admin.indices.create.CreateIndexRequest;
-import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
 import org.elasticsearch.cluster.metadata.IndexMetadata;
 import org.elasticsearch.common.settings.Settings;
 import org.jetbrains.annotations.VisibleForTesting;
@@ -42,7 +39,6 @@ import io.crate.data.Row;
 import io.crate.data.Row1;
 import io.crate.data.RowConsumer;
 import io.crate.execution.ddl.tables.CreateBlobTableRequest;
-import io.crate.execution.ddl.tables.CreateTableResponse;
 import io.crate.execution.ddl.tables.TransportCreateBlobTableAction;
 import io.crate.execution.support.OneRowActionListener;
 import io.crate.expression.symbol.Symbol;
@@ -91,14 +87,11 @@ public class CreateBlobTablePlan implements Plan {
 
         if (plannerContext.clusterState().nodes().getSmallestNonClientNodeVersion().onOrAfter(Version.V_5_10_0)) {
             CreateBlobTableRequest request = new CreateBlobTableRequest(relationName, settings);
-            OneRowActionListener<CreateTableResponse> listener =
-                new OneRowActionListener<>(consumer, ignoredResponse -> new Row1(1L));
-            dependencies.client().execute(TransportCreateBlobTableAction.ACTION, request).whenComplete(listener);
+            dependencies.client().execute(TransportCreateBlobTableAction.ACTION, request)
+                .whenComplete(new OneRowActionListener<>(consumer, _ -> new Row1(1L)));
         } else {
-            CreateIndexRequest createIndexRequest = new CreateIndexRequest(relationName.indexNameOrAlias(), settings);
-            OneRowActionListener<CreateIndexResponse> listener =
-                new OneRowActionListener<>(consumer, ignoredResponse -> new Row1(1L));
-            dependencies.client().execute(CreateIndexAction.INSTANCE, createIndexRequest).whenComplete(listener);
+            throw new UnsupportedOperationException(
+                "Cannot create a blob table in a cluster with mixed 6.0.0 and < 5.10.nodes. All nodes must be >= 5.10");
         }
     }
