--- server/src/main/java/org/elasticsearch/action/admin/indices/create/CreateIndexAction.java
@@ -1,32 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.action.admin.indices.create;
-
-import org.elasticsearch.action.ActionType;
-
-public class CreateIndexAction extends ActionType<CreateIndexResponse> {
-
-    public static final CreateIndexAction INSTANCE = new CreateIndexAction();
-    public static final String NAME = "indices:admin/create";
-
-    private CreateIndexAction() {
-        super(NAME);
-    }
-}
